requirements.txt
apache-airflow
boto3
pandas
seaborn
matplotlib
pyyaml
python-dotenv